+++ gradle.properties
@@ -1,4 +1,4 @@
 release_version = 0.1.0
 
-description = 'th2 data service for fixing wrong status of events'
+description = 'th2 cravler processor for fixing wrong status of events'
 vcs_url=https://github.com/th2-net/th2-crawler-event-healer.git
