+++ src/main/java/com/exactpro/th2/crawler/processor/healer/cfg/HealerConfiguration.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.exactpro.th2.dataservice.healer.cfg;
+package com.exactpro.th2.crawler.processor.healer.cfg;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
